requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.1
PyPDF2>=3.0.1
python-dotenv>=1.0.1
